Another re-implementation for sus_mount and mnt_id_reorder.
With hacking into struct mount* and hooking with fork() and unshare(), we only need to check once on each new process, and mark the value of is_sus_mount and fake_mnt_id for each mount. Hopefully this approch can greatly reduce CPU cycle when reading mountinfo, so as to avoid side channel attack as much as possible.