Another re-implementation for sus_kstat.
With hacking into struct inode in include/linux/fs.h, we only need to open the target file in kernel and update the value of is_sus_mount and sus_kstat. This approch should greatly reduce CPU cycle looping the linked list on each stat syscalls, so as to avoid side channel attack as much as possible.